rgb30: fix swapped volume buttons in keymon.c
My observation shows that you swapped the volume buttons. The left button is volume-up. 
This may also result in the problem, that the volume control is not that reliable and has some jittery behavior. I think you also mentioned this in your TODO.
Maybe its not only because of the swapped button, but because the JELOS (MOSS) base does still handle the buttons itself and keymon interferes with this separate handling.